Added in-game multiplayer logic and fixed disposed resources

+++ RangoUnchained/core/src/main/java/io/github/RangoUnchained/Model/Firebase/Utils/PlayerInLobby.java
@@ -5,7 +5,7 @@ public class PlayerInLobby {
     public String displayName;
     public Boolean isReady;
     public Integer finishScore;
-    public Long finishTime;
+    public Double finishTime;
 
     // Default constructor for Firebase deserialization
     public PlayerInLobby() {
